requirements.txt
streamlit>=1.44,<2
requests>=2.31,<3
python-dotenv>=1.0,<2
datasets>=2.19,<3
fastapi>=0.110,<1
uvicorn>=0.29,<1
